# src/main/java/com/coding/task/rego/io/swagger/api/PersonApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.21).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.coding.task.rego.io.swagger.api;

import com.coding.task.rego.exception.BadRequestException;
import com.coding.task.rego.exception.ResourceNotFoundException;
import com.coding.task.rego.io.swagger.model.ModelApiResponse;
import com.coding.task.rego.io.swagger.model.PersonRequest;
import com.coding.task.rego.io.swagger.model.PersonResponse;
import com.coding.task.rego.model.RegistrationRequest;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import javax.validation.Valid;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2021-07-07T12:54:25.965Z")

@Validated
@Api(value = "person", description = "the person API")
@RequestMapping(value = "/api/v1")
public interface PersonApi {

    @ApiOperation(value = "", nickname = "linkVehicle", notes = "", tags={ "Person", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "successful operation"),
        @ApiResponse(code = 400, message = "Invalid Data"),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/person/{personId}/linkVehicle",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<Void> linkVehicle(@ApiParam(value = "person id",required=true) @PathVariable("personId") Integer personId,@ApiParam(value = "" ,required=true )  @Valid @RequestBody RegistrationRequest item) throws BadRequestException;


    @ApiOperation(value = "", nickname = "personGet", notes = "", response = PersonResponse.class, tags={ "Person", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = PersonResponse.class),
        @ApiResponse(code = 400, message = "Invalid Data"),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/person/{personId}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.GET)
    ResponseEntity<PersonResponse> personGet(@ApiParam(value = "person id",required=true) @PathVariable("personId") Integer personId) throws ResourceNotFoundException;


    @ApiOperation(value = "", nickname = "personPost", notes = "creates a person", response = ModelApiResponse.class, tags={ "Person", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = ModelApiResponse.class),
        @ApiResponse(code = 201, message = "Created"),
        @ApiResponse(code = 400, message = "Invalid Order"),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/person",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<ModelApiResponse> personPost(@ApiParam(value = "" ,required=true )  @Valid @RequestBody PersonRequest item);


    @ApiOperation(value = "", nickname = "unLinkVehicle", notes = "", tags={ "Person", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "successful operation"),
        @ApiResponse(code = 400, message = "Invalid Data"),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/person/{personId}/unLinkVehicle",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<Void> unLinkVehicle(@ApiParam(value = "person id",required=true) @PathVariable("personId") Integer personId,@ApiParam(value = "" ,required=true )  @Valid @RequestBody RegistrationRequest item) throws BadRequestException;

}